Move approval and wizard ownership into application services
Move approval queue, intent cache, wizard state, and their RPC payload helpers behind application-owned modules while keeping existing gateway and sandbox imports as compatibility exports. This rebuilds the slice on current dev because the original cleanup commits depend on later CLI and gateway decompositions that are not part of this batch.

Source-PR: https://github.com/opensquilla/opensquilla/pull/45

Source-PR-Author: lose4578 (Traly)

Source-Commits: 53d4ac1, 209474e, b901b85, e6a1c69, c41850e from integration/architecture-cleanup

Adapted-by: Open-Squilla <275096992+Open-Squilla@users.noreply.github.com>

Constraint: Current dev includes tokenjuice and lacks the intermediate CLI/gateway split files from the cleanup branch.

Rejected: cherry-pick application commits directly | they pull in absent intermediate CLI/gateway files and create modify/delete conflicts.

Confidence: medium

Scope-risk: moderate

Directive: Keep gateway and sandbox compatibility exports until all runtime callers are migrated deliberately.

Tested: PYTHONDONTWRITEBYTECODE=1 PYTHONPATH=src /home/weihe/work/code/rhythm/opensquilla/.venv/bin/pytest -q -p no:cacheprovider tests/test_application/test_approval_rpc.py tests/test_application/test_wizard.py tests/test_gateway/test_approval_queue_persistence.py tests/test_gateway/test_rpc_approvals.py tests/test_gateway/test_rpc_wizard.py tests/test_tools/test_shell_approval_policy.py tests/test_tools/test_apply_patch_gates.py; PYTHONDONTWRITEBYTECODE=1 PYTHONPATH=src /home/weihe/work/code/rhythm/opensquilla/.venv/bin/pytest -q -p no:cacheprovider tests/test_ci/test_architecture_import_contracts.py; git diff --check origin/dev..HEAD

Not-tested: live gateway browser smoke and external channel approval flows
