+++ tests/test_ci/test_architecture_import_contracts.py
@@ -46,6 +46,7 @@ APPROVED_PACKAGE_IMPORTS: frozenset[tuple[str, str]] = frozenset({
     ("engine", "squilla_router"),
     ("engine", "tools"),
     ("gateway", "agents"),
+    ("gateway", "application"),
     ("gateway", "channels"),
     ("gateway", "contracts"),
     ("gateway", "engine"),
@@ -79,6 +80,7 @@ APPROVED_PACKAGE_IMPORTS: frozenset[tuple[str, str]] = frozenset({
     ("onboarding", "provider"),
     ("onboarding", "search"),
     ("provider", "engine"),
+    ("sandbox", "application"),
     ("sandbox", "gateway"),
     ("sandbox", "safety"),
     ("sandbox", "tools"),
